Move unit tests for sub-datasource objects
