+++ shared-libs/cht-datasource/test/index.spec.ts
@@ -2,17 +2,8 @@ import { expect } from 'chai';
 import * as Index from '../src';
 import { DataContext } from '../src';
 import { hasAnyPermission, hasPermissions } from '../src/auth';
-import * as Contact from '../src/contact';
-import * as Person from '../src/person';
-import * as Place from '../src/place';
-import * as Qualifier from '../src/qualifier';
-import * as Report from '../src/report';
-import * as Target from '../src/target';
 import sinon, { SinonStub } from 'sinon';
 import * as Context from '../src/libs/data-context';
-import { Page } from '../src/libs/core';
-import { fakeGenerator } from './utils';
-import { DOC_TYPES } from '@medic/constants';
 
 describe('CHT Script API - getDatasource', () => {
   let dataContext: DataContext;
@@ -67,771 +58,6 @@ describe('CHT Script API - getDatasource', () => {
       expect(result).to.equal(true);
       expect(dataContextBind.calledOnceWithExactly(hasAnyPermission)).to.be.true;
       expect(bound.calledOnceWithExactly([['can_edit']], ['chw'], undefined)).to.be.true;
-    });
-
-    describe('place', () => {
-      let place: typeof v1.place;
-
-      beforeEach(() => place = v1.place);
-
-      it('contains expected keys', () => {
-        expect(place).to.have.all.keys(['getByType', 'getByUuid', 'getByUuidWithLineage', 'getPageByType',
-          'create', 'update']);
-      });
-
-      it('getByUuid', async () => {
-        const expectedPlace = {};
-        const placeGet = sinon.stub().resolves(expectedPlace);
-        dataContextBind.returns(placeGet);
-        const qualifier = { uuid: 'my-places-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedPlace = await place.getByUuid(qualifier.uuid);
-
-        expect(returnedPlace).to.equal(expectedPlace);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.get)).to.be.true;
-        expect(placeGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getByUuidWithLineage', async () => {
-        const expectedPlace = {};
-        const placeGet = sinon.stub().resolves(expectedPlace);
-        dataContextBind.returns(placeGet);
-        const qualifier = { uuid: 'my-places-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedPlace = await place.getByUuidWithLineage(qualifier.uuid);
-
-        expect(returnedPlace).to.equal(expectedPlace);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.getWithLineage)).to.be.true;
-        expect(placeGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getPageByType', async () => {
-        const expectedPlaces: Page<Place.v1.Place> = { data: [], cursor: null };
-        const placeGetPage = sinon.stub().resolves(expectedPlaces);
-        dataContextBind.returns(placeGetPage);
-        const placeType = 'place';
-        const limit = 2;
-        const cursor = '1';
-        const placeTypeQualifier = { contactType: placeType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
-
-        const returnedPlaces = await place.getPageByType(placeType, cursor, limit);
-
-        expect(returnedPlaces).to.equal(expectedPlaces);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.getPage)).to.be.true;
-        expect(placeGetPage.calledOnceWithExactly(placeTypeQualifier, cursor, limit)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(placeType)).to.be.true;
-      });
-
-      it('getPageByType uses default cursor and limit', async () => {
-        const expectedPlaces: Page<Place.v1.Place> = {data: [], cursor: null};
-        const placeGetPage = sinon.stub().resolves(expectedPlaces);
-        dataContextBind.returns(placeGetPage);
-        const placeType = 'place';
-        const placeTypeQualifier = { contactType: placeType };
-        sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
-
-        const returnedPlaces = await place.getPageByType(placeType);
-
-        expect(returnedPlaces).to.equal(expectedPlaces);
-        expect(placeGetPage.calledOnceWithExactly(placeTypeQualifier, null, 100)).to.be.true;
-      });
-
-      it('getByType', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const placeGetAll = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(placeGetAll);
-        const placeType = 'place';
-        const placeTypeQualifier = { contactType: placeType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
-
-        const res = place.getByType(placeType);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.getAll)).to.be.true;
-        expect(placeGetAll.calledOnceWithExactly(placeTypeQualifier)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(placeType)).to.be.true;
-      });
-
-      it('create', async () => {
-        const placeInput = { name: 'p1', type: 'place' };
-        const expectedPlace = {
-          ...placeInput,
-          reported_date: 12312312
-        };
-        const placeCreate = sinon.stub().resolves(expectedPlace);
-        dataContextBind.returns(placeCreate);
-
-        const returnedPlace = await place.create(placeInput);
-
-        expect(returnedPlace).to.equal(expectedPlace);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.create)).to.be.true;
-        expect(placeCreate.calledOnceWithExactly(placeInput)).to.be.true;
-      });
-
-      it('update', async () => {
-        const placeInput = { name: 'p1', type: 'place', _id: '123', _rev: '1-abc' };
-        const expectedPlace = {
-          ...placeInput,
-          reported_date: 12312312
-        };
-        const placeUpdate = sinon.stub().resolves(expectedPlace);
-        dataContextBind.returns(placeUpdate);
-
-        const returnedPlace = await place.update(placeInput);
-
-        expect(returnedPlace).to.equal(expectedPlace);
-        expect(dataContextBind.calledOnceWithExactly(Place.v1.update)).to.be.true;
-        expect(placeUpdate.calledOnceWithExactly(placeInput)).to.be.true;
-      });
-    });
-
-    describe('person', () => {
-      let person: typeof v1.person;
-
-      beforeEach(() => person = v1.person);
-
-      it('contains expected keys', () => {
-        expect(person).to.have.all.keys([
-          'getByType',
-          'getByUuid',
-          'getByUuidWithLineage',
-          'getPageByType',
-          'create',
-          'update'
-        ]);
-      });
-
-      it('getByUuid', async () => {
-        const expectedPerson = {};
-        const personGet = sinon.stub().resolves(expectedPerson);
-        dataContextBind.returns(personGet);
-        const qualifier = { uuid: 'my-persons-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedPerson = await person.getByUuid(qualifier.uuid);
-
-        expect(returnedPerson).to.equal(expectedPerson);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.get)).to.be.true;
-        expect(personGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('create', async () => {
-        const personInput = { name: 'apoorva', type: 'person', parent: 'p1' };
-        const expectedPerson = {
-          ...personInput,
-          reported_date: 12312312
-        };
-        const personCreate = sinon.stub().resolves(expectedPerson);
-        dataContextBind.returns(personCreate);
-
-        const returnedPerson = await person.create(personInput);
-
-        expect(returnedPerson).to.equal(expectedPerson);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.create)).to.be.true;
-        expect(personCreate.calledOnceWithExactly(personInput)).to.be.true;
-      });
-
-      it('update', async () => {
-        const personInput = {
-          name: 'apoorva',
-          type: 'person',
-          parent: { _id: 'p1' },
-          _id: '123',
-          _rev: '1-abc',
-          reported_date: 12312312
-        };
-        const expectedPerson = {
-          ...personInput,
-          _rev: '2-def',
-        };
-        const personUpdate = sinon.stub().resolves(expectedPerson);
-        dataContextBind.returns(personUpdate);
-
-        const returnedPlace = await person.update(personInput);
-
-        expect(returnedPlace).to.equal(expectedPerson);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.update)).to.be.true;
-        expect(personUpdate.calledOnceWithExactly(personInput)).to.be.true;
-      });
-
-      it('getByUuidWithLineage', async () => {
-        const expectedPerson = {};
-        const personGet = sinon.stub().resolves(expectedPerson);
-        dataContextBind.returns(personGet);
-        const qualifier = { uuid: 'my-persons-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedPerson = await person.getByUuidWithLineage(qualifier.uuid);
-
-        expect(returnedPerson).to.equal(expectedPerson);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.getWithLineage)).to.be.true;
-        expect(personGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getPageByType', async () => {
-        const expectedPeople: Page<Person.v1.Person> = { data: [], cursor: null };
-        const personGetPage = sinon.stub().resolves(expectedPeople);
-        dataContextBind.returns(personGetPage);
-        const personType = 'person';
-        const limit = 2;
-        const cursor = '1';
-        const personTypeQualifier = { contactType: personType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(personTypeQualifier);
-
-        const returnedPeople = await person.getPageByType(personType, cursor, limit);
-
-        expect(returnedPeople).to.equal(expectedPeople);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.getPage)).to.be.true;
-        expect(personGetPage.calledOnceWithExactly(personTypeQualifier, cursor, limit)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(personType)).to.be.true;
-      });
-
-      it('getPageByType uses default cursor and limit', async () => {
-        const expectedPeople: Page<Person.v1.Person> = {data: [], cursor: null};
-        const personGetPage = sinon.stub().resolves(expectedPeople);
-        dataContextBind.returns(personGetPage);
-        const personType = 'person';
-        const personTypeQualifier = { contactType: personType };
-        sinon.stub(Qualifier, 'byContactType').returns(personTypeQualifier);
-
-        const returnedPeople = await person.getPageByType(personType);
-
-        expect(returnedPeople).to.equal(expectedPeople);
-        expect(personGetPage.calledOnceWithExactly(personTypeQualifier, null, 100)).to.be.true;
-      });
-
-      it('getByType', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const personGetAll = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(personGetAll);
-        const personType = 'person';
-        const personTypeQualifier = { contactType: personType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(personTypeQualifier);
-
-        const res = person.getByType(personType);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Person.v1.getAll)).to.be.true;
-        expect(personGetAll.calledOnceWithExactly(personTypeQualifier)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(personType)).to.be.true;
-      });
-    });
-
-    describe('contact', () => {
-      let contact: typeof v1.contact;
-
-      beforeEach(() => contact = v1.contact);
-
-      it('contains expected keys', () => {
-        expect(contact).to.have.all.keys(
-          [
-            'getByUuid',
-            'getByUuidWithLineage',
-            'getUuidsByTypeFreetext',
-            'getUuidsPageByTypeFreetext',
-            'getUuidsPageByFreetext',
-            'getUuidsByFreetext',
-            'getUuidsPageByType',
-            'getUuidsByType',
-          ]
-        );
-      });
-
-      it('getByUuid', async () => {
-        const expectedContact = {};
-        const contactGet = sinon.stub().resolves(expectedContact);
-        dataContextBind.returns(contactGet);
-        const qualifier = { uuid: 'my-contact-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedContact = await contact.getByUuid(qualifier.uuid);
-
-        expect(returnedContact).to.equal(expectedContact);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.get)).to.be.true;
-        expect(contactGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getByUuidWithLineage', async () => {
-        const expectedContact = {};
-        const contactGet = sinon.stub().resolves(expectedContact);
-        dataContextBind.returns(contactGet);
-        const qualifier = { uuid: 'my-contact-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedContact = await contact.getByUuidWithLineage(qualifier.uuid);
-
-        expect(returnedContact).to.equal(expectedContact);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getWithLineage)).to.be.true;
-        expect(contactGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getUuidsPageByTypeFreetext uses default cursor and limit', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = {data: [], cursor: null};
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const freetext = 'abc';
-        const contactType = 'person';
-        const qualifier = { contactType, freetext };
-        sinon.stub(Qualifier, 'and').returns(qualifier);
-        sinon.stub(Qualifier, 'byFreetext').returns({ freetext });
-        sinon.stub(Qualifier, 'byContactType').returns({ contactType });
-
-        const returnedContactIds = await contact.getUuidsPageByTypeFreetext(freetext, contactType);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(contactGetIdsPage.calledOnceWithExactly(qualifier, null, 10000)).to.be.true;
-      });
-
-      it('getUuidsPageByTypeFreetext', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = { data: [], cursor: null };
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const freetext = 'abc';
-        const contactType = 'person';
-        const limit = 2;
-        const cursor = '1';
-        const contactTypeQualifier = { contactType };
-        const freetextQualifier = { freetext };
-        const qualifier = { contactType, freetext };
-        const andQualifier = sinon.stub(Qualifier, 'and').returns(qualifier);
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(freetextQualifier);
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(contactTypeQualifier);
-
-        const returnedContactIds = await contact.getUuidsPageByTypeFreetext(freetext, contactType, cursor, limit);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuidsPage)).to.be.true;
-        expect(
-          contactGetIdsPage.calledOnceWithExactly(qualifier, cursor, limit)
-        ).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(contactType)).to.be.true;
-        expect(andQualifier.calledOnceWithExactly(freetextQualifier, contactTypeQualifier)).to.be.true;
-      });
-
-      it('getUuidsPageByType', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = { data: [], cursor: null };
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const contactType = 'person';
-        const limit = 2;
-        const cursor = '1';
-        const contactTypeQualifier = { contactType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(contactTypeQualifier);
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext');
-
-        const returnedContactIds = await contact.getUuidsPageByType(contactType, cursor, limit);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuidsPage)).to.be.true;
-        expect(
-          contactGetIdsPage.calledOnceWithExactly(contactTypeQualifier, cursor, limit)
-        ).to.be.true;
-        expect(byContactType.calledOnceWithExactly(contactType)).to.be.true;
-        expect(byFreetext.notCalled).to.be.true;
-      });
-
-      it('getUuidsPageByType uses default cursor and limit', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = {data: [], cursor: null};
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const contactType = 'person';
-        const contactTypeQualifier = { contactType };
-        sinon.stub(Qualifier, 'byContactType').returns(contactTypeQualifier);
-
-        const returnedContactIds = await contact.getUuidsPageByType(contactType);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(contactGetIdsPage.calledOnceWithExactly(contactTypeQualifier, null, 10000)).to.be.true;
-      });
-
-      it('getUuidsPageByFreetext', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = { data: [], cursor: null };
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const freetext = 'abc';
-        const limit = 2;
-        const cursor = '1';
-        const freetextQualifier = { freetext };
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(freetextQualifier);
-        const byContactType = sinon.stub(Qualifier, 'byContactType');
-
-        const returnedContactIds = await contact.getUuidsPageByFreetext(freetext, cursor, limit);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuidsPage)).to.be.true;
-        expect(
-          contactGetIdsPage.calledOnceWithExactly(freetextQualifier, cursor, limit)
-        ).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-        expect(byContactType.notCalled).to.be.true;
-      });
-
-      it('getUuidsPageByFreetext uses default cursor and limit', async () => {
-        const expectedContactIds: Page<Contact.v1.Contact> = {data: [], cursor: null};
-        const contactGetIdsPage = sinon.stub().resolves(expectedContactIds);
-        dataContextBind.returns(contactGetIdsPage);
-        const freetext = 'abc';
-        const freetextQualifier = { freetext };
-        sinon.stub(Qualifier, 'byFreetext').returns(freetextQualifier);
-
-        const returnedContactIds = await contact.getUuidsPageByFreetext(freetext);
-
-        expect(returnedContactIds).to.equal(expectedContactIds);
-        expect(contactGetIdsPage.calledOnceWithExactly(freetextQualifier, null, 10000)).to.be.true;
-      });
-
-      it('getUuidsByTypeFreetext', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const contactGetIds = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(contactGetIds);
-        const freetext = 'abc';
-        const contactType = 'person';
-        const contactTypeQualifier = { contactType };
-        const freetextQualifier = { freetext };
-        const qualifier = { contactType, freetext };
-        const andQualifier = sinon.stub(Qualifier, 'and').returns(qualifier);
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(freetextQualifier);
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(contactTypeQualifier);
-
-        const res = contact.getUuidsByTypeFreetext(freetext, contactType);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuids)).to.be.true;
-        expect(contactGetIds.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(andQualifier.calledOnceWithExactly(freetextQualifier, contactTypeQualifier)).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(contactType)).to.be.true;
-      });
-
-      it('getUuidsByType', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const contactGetIds = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(contactGetIds);
-        const contactType = 'person';
-        const contactTypeQualifier = { contactType };
-        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(contactTypeQualifier);
-
-        const res = contact.getUuidsByType(contactType);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuids)).to.be.true;
-        expect(contactGetIds.calledOnceWithExactly(contactTypeQualifier)).to.be.true;
-        expect(byContactType.calledOnceWithExactly(contactType)).to.be.true;
-      });
-
-      it('getUuidsByFreetext', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const contactGetIds = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(contactGetIds);
-        const freetext = 'abc';
-        const freetextQualifier = { freetext };
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(freetextQualifier);
-
-        const res = contact.getUuidsByFreetext(freetext);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Contact.v1.getUuids)).to.be.true;
-        expect(contactGetIds.calledOnceWithExactly(freetextQualifier)).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-      });
-    });
-
-    describe('report', () => {
-      let report: typeof v1.report;
-
-      beforeEach(() => report = v1.report);
-
-      it('contains expected keys', () => {
-        expect(report).to.have.all.keys([
-          'getUuidsByFreetext',
-          'getUuidsPageByFreetext',
-          'getByUuid',
-          'create',
-          'update',
-          'getByUuidWithLineage',
-        ]);
-      });
-
-      it('getByUuid', async () => {
-        const expectedReport = {};
-        const reportGet = sinon.stub().resolves(expectedReport);
-        dataContextBind.returns(reportGet);
-        const qualifier = { uuid: 'my-report-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedReport = await report.getByUuid(qualifier.uuid);
-
-        expect(returnedReport).to.equal(expectedReport);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.get)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getByUuidWithLineage', async () => {
-        const expectedReport = {};
-        const reportGet = sinon.stub().resolves(expectedReport);
-        dataContextBind.returns(reportGet);
-        const qualifier = { uuid: 'my-report-uuid' };
-        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
-
-        const returnedReport = await report.getByUuidWithLineage(qualifier.uuid);
-
-        expect(returnedReport).to.equal(expectedReport);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.getWithLineage)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
-      });
-
-      it('getUuidsPageByFreetext', async () => {
-        const expectedReportIds: Page<Report.v1.Report> = { data: [], cursor: null };
-        const reportGetIdsPage = sinon.stub().resolves(expectedReportIds);
-        dataContextBind.returns(reportGetIdsPage);
-        const freetext = 'abc';
-        const limit = 2;
-        const cursor = '1';
-        const qualifier = { freetext };
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(qualifier);
-
-        const returnedContactIds = await report.getUuidsPageByFreetext(freetext, cursor, limit);
-
-        expect(returnedContactIds).to.equal(expectedReportIds);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.getUuidsPage)).to.be.true;
-        expect(
-          reportGetIdsPage.calledOnceWithExactly(qualifier, cursor, limit)
-        ).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-      });
-
-      it('getUuidsPageByFreetext uses default cursor and limit', async () => {
-        const expectedReportIds: Page<Report.v1.Report> = {data: [], cursor: null};
-        const reportGetIdsPage = sinon.stub().resolves(expectedReportIds);
-        dataContextBind.returns(reportGetIdsPage);
-        const freetext = 'abc';
-        const qualifier = { freetext };
-        sinon.stub(Qualifier, 'byFreetext').returns(qualifier);
-
-        const returnedContactIds = await report.getUuidsPageByFreetext(freetext);
-
-        expect(returnedContactIds).to.equal(expectedReportIds);
-        expect(reportGetIdsPage.calledOnceWithExactly(qualifier, null, 10000)).to.be.true;
-      });
-
-      it('getUuidsByFreetext', () => {
-        const mockAsyncGenerator = fakeGenerator();
-
-        const contactGetIds = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(contactGetIds);
-        const freetext = 'abc';
-        const qualifier = { freetext };
-        const byFreetext = sinon.stub(Qualifier, 'byFreetext').returns(qualifier);
-
-        const res = report.getUuidsByFreetext(freetext);
-
-        expect(res).to.deep.equal(mockAsyncGenerator);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.getUuids)).to.be.true;
-        expect(contactGetIds.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(byFreetext.calledOnceWithExactly(freetext)).to.be.true;
-      });
-
-      it('create', async () => {
-        const reportInput = { form: 'apoorva', type: 'report', contact: 'c1' };
-        const expectedReport = {
-          ...reportInput,
-          reported_date: 12312312
-        };
-        const reportCreate = sinon.stub().resolves(expectedReport);
-        dataContextBind.returns(reportCreate);
-
-        const returnedReport = await report.create(reportInput);
-
-        expect(returnedReport).to.equal(expectedReport);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.create)).to.be.true;
-        expect(reportCreate.calledOnceWithExactly(reportInput)).to.be.true;
-      });
-
-      it('update', async () => {
-        const reportInput = {
-          form: 'apoorva',
-          type: DOC_TYPES.DATA_RECORD,
-          contact: { _id: 'c1' },
-          _id: '123',
-          _rev: '1-abc',
-          reported_date: 12312312,
-          fields: {}
-        };
-        const expectedReport = {
-          ...reportInput,
-          _rev: '2-def',
-        };
-        const reportUpdate = sinon.stub().resolves(expectedReport);
-        dataContextBind.returns(reportUpdate);
-
-        const returnedPlace = await report.update(reportInput);
-
-        expect(returnedPlace).to.equal(expectedReport);
-        expect(dataContextBind.calledOnceWithExactly(Report.v1.update)).to.be.true;
-        expect(reportUpdate.calledOnceWithExactly(reportInput)).to.be.true;
-      });
-    });
-
-    describe('target', () => {
-      let target: typeof v1.target;
-
-      beforeEach(() => target = v1.target);
-
-      it('contains expected keys', () => {
-        expect(target).to.have.all.keys([
-          'getById', 'getByReportingPeriodContactIdUsername',
-          'getPageByReportingPeriodContactIds', 'getByReportingPeriodContactIds'
-        ]);
-      });
-
-      it('getById', async () => {
-        const expectedTarget = {};
-        const reportGet = sinon.stub().resolves(expectedTarget);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.byId('my-target-uuid');
-
-        const returnedTarget = await target.getById(qualifier.id);
-
-        expect(returnedTarget).to.equal(expectedTarget);
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.get)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-      });
-
-      it('getByReportingPeriodContactIdUsername', async () => {
-        const expectedTarget = {};
-        const reportGet = sinon.stub().resolves(expectedTarget);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01' ),
-          Qualifier.byContactId('my-contact-uuid'),
-          Qualifier.byUsername('my-username')
-        );
-
-        const returnedTarget = await target.getByReportingPeriodContactIdUsername(
-          qualifier.reportingPeriod,
-          qualifier.contactId,
-          qualifier.username
-        );
-
-        expect(returnedTarget).to.equal(expectedTarget);
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.get)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-      });
-
-      it('getPageByReportingPeriodContactIds uses default cursor and limit', async () => {
-        const expectedTarget = {};
-        const reportGet = sinon.stub().resolves(expectedTarget);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01'),
-          Qualifier.byContactIds(['my-first-contact-uuid', 'my-second-contact-uuid'])
-        );
-
-        const returnedTarget = await target.getPageByReportingPeriodContactIds(
-          qualifier.reportingPeriod,
-          qualifier.contactIds
-        );
-
-        expect(returnedTarget).to.equal(expectedTarget);
-        expect(reportGet.calledOnceWithExactly(qualifier, null, 100)).to.be.true;
-      });
-
-      it('getPageByReportingPeriodContactIds - multiple contact Ids', async () => {
-        const expectedTarget = {};
-        const reportGet = sinon.stub().resolves(expectedTarget);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01'),
-          Qualifier.byContactIds(['my-first-contact-uuid', 'my-second-contact-uuid'])
-        );
-
-        const returnedTarget = await target.getPageByReportingPeriodContactIds(
-          qualifier.reportingPeriod,
-          qualifier.contactIds,
-          '1',
-          10
-        );
-
-        expect(returnedTarget).to.equal(expectedTarget);
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.getPage)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier, '1', 10)).to.be.true; 
-      });
-
-      it('getPageByReportingPeriodContactIds - since contact Id', async () => {
-        const expectedTarget = {};
-        const reportGet = sinon.stub().resolves(expectedTarget);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01'),
-          Qualifier.byContactId('my-first-contact-uuid')
-        );
-
-        const returnedTarget = await target.getPageByReportingPeriodContactIds(
-          qualifier.reportingPeriod,
-          qualifier.contactId,
-          '1',
-          10
-        );
-
-        expect(returnedTarget).to.equal(expectedTarget);
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.getPage)).to.be.true;
-        expect(reportGet.calledOnceWithExactly(qualifier, '1', 10)).to.be.true;
-      });
-      
-      it('getByReportingPeriodContactIds multiple contact Ids', () => {
-        const mockAsyncGenerator = fakeGenerator();
-        const reportGet = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01'),
-          Qualifier.byContactIds(['my-first-contact-uuid', 'my-second-contact-uuid'])
-        );
-        
-        const returnedTarget = target.getByReportingPeriodContactIds(
-          qualifier.reportingPeriod,
-          qualifier.contactIds
-        );
-        
-        expect(returnedTarget).to.deep.equal(mockAsyncGenerator);
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.getAll)).to.be.true;
-      });
-
-      it('getByReportingPeriodContactIds since contact Id', () => {
-        const mockAsyncGenerator = fakeGenerator();
-        const reportGet = sinon.stub().returns(mockAsyncGenerator);
-        dataContextBind.returns(reportGet);
-        const qualifier = Qualifier.and(
-          Qualifier.byReportingPeriod('2020-01'),
-          Qualifier.byContactId('my-first-contact-uuid')
-        );
-
-        const returnedTarget = target.getByReportingPeriodContactIds(
-          qualifier.reportingPeriod,
-          qualifier.contactId
-        );
-
-        expect(returnedTarget).to.deep.equal(mockAsyncGenerator);
-        expect(reportGet.calledOnceWithExactly(qualifier)).to.be.true;
-        expect(dataContextBind.calledOnceWithExactly(Target.v1.getAll)).to.be.true;
-      });
     });
   });
 });
