+++ shared-libs/cht-datasource/test/place.spec.ts
@@ -6,18 +6,20 @@ import * as Input from '../src/input';
 import * as Context from '../src/libs/data-context';
 import sinon, { SinonStub } from 'sinon';
 import { expect } from 'chai';
-import { DataContext } from '../src';
+import { DataContext, Page } from '../src';
 import * as Core from '../src/libs/core';
 import { fakeGenerator } from './utils';
 
 describe('place', () => {
-  const dataContext = { } as DataContext;
+  const dataContext = { bind: () => null } as DataContext;
+  let dataContextBind: SinonStub;
   let assertDataContext: SinonStub;
   let adapt: SinonStub;
   let isUuidQualifier: SinonStub;
   let isContactTypeQualifier: SinonStub;
 
   beforeEach(() => {
+    dataContextBind = sinon.stub(dataContext, 'bind');
     assertDataContext = sinon.stub(Context, 'assertDataContext');
     adapt = sinon.stub(Context, 'adapt');
     isUuidQualifier = sinon.stub(Qualifier, 'isUuidQualifier');
@@ -400,6 +402,129 @@ describe('place', () => {
         expect(adapt.calledOnceWithExactly(dataContext, Local.Place.v1.update, Remote.Place.v1.update))
           .to.be.true;
         expect(updatePlaceDoc.notCalled).to.be.true;
+      });
+    });
+
+    describe('getDatasource', () => {
+      let place: Place.v1.Datasource;
+
+      beforeEach(() => place = Place.v1.getDatasource(dataContext));
+
+      it('contains expected keys', () => {
+        expect(place).to.have.all.keys([
+          'getByType', 'getByUuid', 'getByUuidWithLineage', 'getPageByType', 'create', 'update'
+        ]);
+      });
+
+      it('getByUuid', async () => {
+        const expectedPlace = {};
+        const placeGet = sinon.stub().resolves(expectedPlace);
+        dataContextBind.returns(placeGet);
+        const qualifier = { uuid: 'my-places-uuid' };
+        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
+
+        const returnedPlace = await place.getByUuid(qualifier.uuid);
+
+        expect(returnedPlace).to.equal(expectedPlace);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.get)).to.be.true;
+        expect(placeGet.calledOnceWithExactly(qualifier)).to.be.true;
+        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
+      });
+
+      it('getByUuidWithLineage', async () => {
+        const expectedPlace = {};
+        const placeGet = sinon.stub().resolves(expectedPlace);
+        dataContextBind.returns(placeGet);
+        const qualifier = { uuid: 'my-places-uuid' };
+        const byUuid = sinon.stub(Qualifier, 'byUuid').returns(qualifier);
+
+        const returnedPlace = await place.getByUuidWithLineage(qualifier.uuid);
+
+        expect(returnedPlace).to.equal(expectedPlace);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.getWithLineage)).to.be.true;
+        expect(placeGet.calledOnceWithExactly(qualifier)).to.be.true;
+        expect(byUuid.calledOnceWithExactly(qualifier.uuid)).to.be.true;
+      });
+
+      it('getPageByType', async () => {
+        const expectedPlaces: Page<Place.v1.Place> = { data: [], cursor: null };
+        const placeGetPage = sinon.stub().resolves(expectedPlaces);
+        dataContextBind.returns(placeGetPage);
+        const placeType = 'place';
+        const limit = 2;
+        const cursor = '1';
+        const placeTypeQualifier = { contactType: placeType };
+        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
+
+        const returnedPlaces = await place.getPageByType(placeType, cursor, limit);
+
+        expect(returnedPlaces).to.equal(expectedPlaces);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.getPage)).to.be.true;
+        expect(placeGetPage.calledOnceWithExactly(placeTypeQualifier, cursor, limit)).to.be.true;
+        expect(byContactType.calledOnceWithExactly(placeType)).to.be.true;
+      });
+
+      it('getPageByType uses default cursor and limit', async () => {
+        const expectedPlaces: Page<Place.v1.Place> = {data: [], cursor: null};
+        const placeGetPage = sinon.stub().resolves(expectedPlaces);
+        dataContextBind.returns(placeGetPage);
+        const placeType = 'place';
+        const placeTypeQualifier = { contactType: placeType };
+        sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
+
+        const returnedPlaces = await place.getPageByType(placeType);
+
+        expect(returnedPlaces).to.equal(expectedPlaces);
+        expect(placeGetPage.calledOnceWithExactly(placeTypeQualifier, null, 100)).to.be.true;
+      });
+
+      it('getByType', () => {
+        const mockAsyncGenerator = fakeGenerator();
+
+        const placeGetAll = sinon.stub().returns(mockAsyncGenerator);
+        dataContextBind.returns(placeGetAll);
+        const placeType = 'place';
+        const placeTypeQualifier = { contactType: placeType };
+        const byContactType = sinon.stub(Qualifier, 'byContactType').returns(placeTypeQualifier);
+
+        const res = place.getByType(placeType);
+
+        expect(res).to.deep.equal(mockAsyncGenerator);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.getAll)).to.be.true;
+        expect(placeGetAll.calledOnceWithExactly(placeTypeQualifier)).to.be.true;
+        expect(byContactType.calledOnceWithExactly(placeType)).to.be.true;
+      });
+
+      it('create', async () => {
+        const placeInput = { name: 'p1', type: 'place' };
+        const expectedPlace = {
+          ...placeInput,
+          reported_date: 12312312
+        };
+        const placeCreate = sinon.stub().resolves(expectedPlace);
+        dataContextBind.returns(placeCreate);
+
+        const returnedPlace = await place.create(placeInput);
+
+        expect(returnedPlace).to.equal(expectedPlace);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.create)).to.be.true;
+        expect(placeCreate.calledOnceWithExactly(placeInput)).to.be.true;
+      });
+
+      it('update', async () => {
+        const placeInput = { name: 'p1', type: 'place', _id: '123', _rev: '1-abc' };
+        const expectedPlace = {
+          ...placeInput,
+          reported_date: 12312312
+        };
+        const placeUpdate = sinon.stub().resolves(expectedPlace);
+        dataContextBind.returns(placeUpdate);
+
+        const returnedPlace = await place.update(placeInput);
+
+        expect(returnedPlace).to.equal(expectedPlace);
+        expect(dataContextBind.calledOnceWithExactly(Place.v1.update)).to.be.true;
+        expect(placeUpdate.calledOnceWithExactly(placeInput)).to.be.true;
       });
     });
   });
